Moved knockback up in class tree
